requirements.txt
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24
google-api-python-client==2.118.0
tavily-python>=0.3.0
streamlit==1.31.0
python-dotenv==1.0.1
pydantic==2.6.1
tiktoken>=0.5.2